keep only code and data related to the gps trace match example
