--- gers/examples/python/data/macon-points.csv
@@ -1,6 +1,0 @@
-wkt,payload
-POINT (-83.6271417 32.8519936),payload1
-POINT (-83.6141122 32.8375626),payload2
-POINT (-83.6308682 32.8338996),payload3
-POINT (-83.6312115 32.8334443),payload4
-POINT (-83.6708248 32.8370307),payload5
